// src/enums/shootingMode.ts
export enum ShootingMode {
  Auto = "AUTO", // Automatický režim
  Program = "P", // Program AE (automatická expozice s možností ručních úprav)
  AperturePriority = "Priorita clony",
  ShutterPriority = "Priorita času",
  Manual = "M", // Manuální režim
  Bulb = "B", // Bulb režim (dlouhá expozice)
  Scene = "SCN", // Scénický režim (např. portrét, krajina)
  Panorama = "PANORAMA", // Panorama režim
  Sports = "SPORT", // Sportovní režim (rychlá závěrka)
  Portrait = "PORTRAIT", // Portrét
  Night = "NIGHT", // Noční režim
  Macro = "MACRO", // Makro režim
}